[Test code] Shrink the sizes of some values
This commit changes several types in the test code to use smaller types.
This means we don't have to downcast them to use them, and we can safely
convert them to larger types when needed.

+++ src/vmm/src/devices/virtio/balloon/test_utils.rs
@@ -27,18 +27,18 @@ pub fn invoke_handler_for_queue_event(b: &mut Balloon, queue_index: usize) {
     assert!(b.irq_trigger.has_pending_irq(IrqType::Vring));
 }
 
-pub fn set_request(queue: &VirtQueue, idx: usize, addr: u64, len: u32, flags: u16) {
+pub fn set_request(queue: &VirtQueue, idx: u16, addr: u64, len: u32, flags: u16) {
     // Set the index of the next request.
-    queue.avail.idx.set((idx + 1) as u16);
+    queue.avail.idx.set(idx + 1);
     // Set the current descriptor table entry index.
-    queue.avail.ring[idx].set(idx as u16);
+    queue.avail.ring[idx as usize].set(idx);
     // Set the current descriptor table entry.
-    queue.dtable[idx].set(addr, len, flags, 1);
+    queue.dtable[idx as usize].set(addr, len, flags, 1);
 }
 
 pub fn check_request_completion(queue: &VirtQueue, idx: usize) {
     // Check that the next used will be idx + 1.
-    assert_eq!(queue.used.idx.get(), (idx + 1) as u16);
+    assert_eq!(queue.used.idx.get() as usize, idx + 1);
     // Check that the current used is idx.
     assert_eq!(queue.used.ring[idx].get().id, idx as u32);
     // The length of the completed request is 0.
